@@ -2,16 +2,21 @@ package cc.rits.membership.console.iam.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.authentication.AuthenticationProvider;
+import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityCustomizer;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.oauth2.jwt.JwtDecoder;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.context.DelegatingSecurityContextRepository;
+import org.springframework.security.web.context.HttpSessionSecurityContextRepository;
+import org.springframework.security.web.context.RequestAttributeSecurityContextRepository;
+import org.springframework.security.web.context.SecurityContextRepository;
 
 import cc.rits.membership.console.iam.config.auth.ClientAuthenticationEntryPoint;
-import cc.rits.membership.console.iam.config.auth.IamAuthenticationProvider;
-import cc.rits.membership.console.iam.config.auth.IamUserDetailsService;
+import cc.rits.membership.console.iam.config.auth.CustomUserDetailsService;
+import cc.rits.membership.console.iam.config.auth.UserAccessDeniedHandler;
 import cc.rits.membership.console.iam.config.auth.UserAuthenticationEntryPoint;
 import lombok.RequiredArgsConstructor;
 
@@ -25,18 +30,15 @@ public class WebSecurityConfig {
 
     private final UserAuthenticationEntryPoint userAuthenticationEntryPoint;
 
+    private final UserAccessDeniedHandler userAccessDeniedHandler;
+
     private final ClientAuthenticationEntryPoint clientAuthenticationEntryPoint;
 
-    private final IamUserDetailsService userDetailsService;
+    private final CustomUserDetailsService userDetailsService;
 
     private final PasswordEncoder passwordEncoder;
 
     private final JwtDecoder jwtDecoder;
-
-    @Bean
-    public WebSecurityCustomizer webSecurityCustomizer() {
-        return (web) -> web.ignoring().antMatchers("/v3/api-docs/**", "/swagger-resources/**", "/swagger-ui/**", "/webjars/**");
-    }
 
     @Bean
     public SecurityFilterChain filterChain(final HttpSecurity http) throws Exception {
@@ -47,13 +49,17 @@ public class WebSecurityConfig {
         http.csrf().disable();
 
         // アクセス許可
-        http.authorizeRequests() //
-            .antMatchers("/api/health", "/api/login", "/api/request_password_reset", "/api/password_reset").permitAll() //
-            .antMatchers("/api/oauth2/**").permitAll() //
-            .antMatchers("/api/**").hasRole("USER") //
-            .antMatchers("/**").permitAll() //
-            .anyRequest().authenticated() //
-            .and().exceptionHandling().authenticationEntryPoint(this.userAuthenticationEntryPoint);
+        http.authorizeHttpRequests() //
+            // REST API
+            .requestMatchers("/api/health", "/api/login", "/api/request_password_reset", "/api/password_reset").permitAll() //
+            .requestMatchers("/api/oauth2/**").permitAll() //
+            .requestMatchers("/api/**").hasRole("USER") //
+            // 静的コンテンツ
+            .requestMatchers("/v3/api-docs/**", "/swagger-resources/**", "/swagger-ui/**", "/webjars/**").permitAll() //
+            .anyRequest().authenticated();
+        http.exceptionHandling() //
+            .authenticationEntryPoint(this.userAuthenticationEntryPoint) //
+            .accessDeniedHandler(this.userAccessDeniedHandler);
 
         // リソースサーバ
         http.oauth2ResourceServer() //
@@ -65,11 +71,17 @@ public class WebSecurityConfig {
     }
 
     @Bean
-    public IamAuthenticationProvider oauth2AuthenticationProvider() {
-        final var authenticationProvider = new IamAuthenticationProvider();
+    public AuthenticationProvider authenticationProvider() {
+        final var authenticationProvider = new DaoAuthenticationProvider();
         authenticationProvider.setUserDetailsService(this.userDetailsService);
         authenticationProvider.setPasswordEncoder(this.passwordEncoder);
         return authenticationProvider;
+    }
+
+    @Bean
+    public SecurityContextRepository securityContextRepository() {
+        return new DelegatingSecurityContextRepository(new RequestAttributeSecurityContextRepository(),
+            new HttpSessionSecurityContextRepository());
     }
 
 }
