@@ -1,16 +1,17 @@
 package cc.rits.membership.console.iam.usecase.front.user;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpSession;
-
+import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.web.context.SecurityContextRepository;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
-import cc.rits.membership.console.iam.config.auth.IamAuthenticationProvider;
 import cc.rits.membership.console.iam.infrastructure.api.request.LoginRequest;
 import cc.rits.membership.console.iam.property.AuthProperty;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 
 /**
@@ -20,11 +21,15 @@ import lombok.RequiredArgsConstructor;
 @Component
 public class LoginUseCase {
 
-    private final IamAuthenticationProvider authenticationProvider;
+    private final AuthenticationProvider authenticationProvider;
 
-    private final HttpSession httpSession;
+    private final SecurityContextRepository securityContextRepository;
 
     private final HttpServletRequest httpServletRequest;
+
+    private final HttpServletResponse httpServletResponse;
+
+    private final HttpSession httpSession;
 
     private final AuthProperty authProperty;
 
@@ -42,6 +47,7 @@ public class LoginUseCase {
         SecurityContextHolder.getContext().setAuthentication(authentication);
         this.httpSession.setMaxInactiveInterval(this.authProperty.getSessionTimeout());
         this.httpServletRequest.changeSessionId();
+        this.securityContextRepository.saveContext(SecurityContextHolder.getContext(), this.httpServletRequest, this.httpServletResponse);
     }
 
 }
