@@ -1,11 +1,10 @@
 package cc.rits.membership.console.iam.usecase.front.user;
 
-import javax.servlet.http.HttpSession;
-
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 
 /**
